requirements.txt
flask==3.0.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
ezdxf==1.3.3
openai==1.56.0
python-dotenv==1.0.1
trimesh==4.4.3
shapely==2.0.4
mapbox-earcut==1.0.1
gunicorn==22.0.0
